backend/requirements.txt
gunicorn==20.1.0 
Django==3.2.16
djoser==2.1.0
django-colorfield==0.11.0
django-filter==23.3
django-import-export==3.3.6
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
drf-extra-fields==3.7.0
flake8==6.0.0
flake8-isort==6.0.0
Pillow==10.1.0
psycopg2-binary==2.9.3
python-dotenv==1.0.0
reportlab==4.0.7
